training/flux/train_instruct_pix2pix_flux.py
"""
Flux InstructPix2Pix Training Script
Based on Flux architecture with conditional input image + text instruction
Uses LoRA for efficient fine-tuning on 8xA100

Note: This is a custom implementation as Flux is very new.
      For production use, check for official scripts at:
      https://github.com/black-forest-labs/flux
      https://github.com/huggingface/diffusers
"""

import argparse
import logging
import math
import os
import shutil
from pathlib import Path

import torch
import torch.nn.functional as F
import transformers
from accelerate import Accelerator
from accelerate.logging import get_logger
from accelerate.utils import ProjectConfiguration, set_seed
from datasets import load_from_disk
from peft import LoraConfig, get_peft_model
from tqdm.auto import tqdm
from transformers import CLIPTextModel, CLIPTokenizer, T5EncoderModel, T5TokenizerFast

import diffusers
from diffusers import FlowMatchEulerDiscreteScheduler, AutoencoderKL
from diffusers.optimization import get_scheduler
from diffusers.utils import check_min_version

# Note: Flux uses different components than SD/SDXL
# This is a simplified version - for production, use official scripts when available

logger = get_logger(__name__, log_level="INFO")

def parse_args():
    parser = argparse.ArgumentParser(description="Train Flux for InstructPix2Pix")
    parser.add_argument(
        "--pretrained_model_name_or_path",
        type=str,
        default="black-forest-labs/FLUX.1-dev",
        help="Path to pretrained Flux model",
    )
    parser.add_argument(
        "--dataset_name",
        type=str,
        required=True,
        help="Path to HuggingFace dataset",
    )
    parser.add_argument(
        "--output_dir",
        type=str,
        default="./output/instruct-pix2pix-flux",
        help="Output directory for checkpoints",
    )
    parser.add_argument(
        "--resolution",
        type=int,
        default=1024,
        help="Image resolution",
    )
    parser.add_argument(
        "--train_batch_size",
        type=int,
        default=2,
        help="Batch size per GPU (Flux is large!)",
    )
    parser.add_argument(
        "--num_train_epochs",
        type=int,
        default=100,
        help="Number of training epochs",
    )
    parser.add_argument(
        "--learning_rate",
        type=float,
        default=1e-4,
        help="Learning rate",
    )
    parser.add_argument(
        "--rank",
        type=int,
        default=128,
        help="LoRA rank",
    )
    parser.add_argument(
        "--seed",
        type=int,
        default=42,
        help="Random seed",
    )
    parser.add_argument(
        "--checkpointing_steps",
        type=int,
        default=500,
        help="Save checkpoint every N steps",
    )
    parser.add_argument(
        "--gradient_accumulation_steps",
        type=int,
        default=2,
        help="Gradient accumulation steps",
    )
    parser.add_argument(
        "--mixed_precision",
        type=str,
        default="bf16",
        choices=["no", "fp16", "bf16"],
        help="Mixed precision training",
    )
    parser.add_argument(
        "--report_to",
        type=str,
        default="wandb",
        help="Logging backend",
    )
    
    args = parser.parse_args()
    return args


def main():
    args = parse_args()
    
    # Setup accelerator
    accelerator_project_config = ProjectConfiguration(
        project_dir=args.output_dir,
        logging_dir=Path(args.output_dir) / "logs",
    )
    
    accelerator = Accelerator(
        gradient_accumulation_steps=args.gradient_accumulation_steps,
        mixed_precision=args.mixed_precision,
        log_with=args.report_to,
        project_config=accelerator_project_config,
    )
    
    # Setup logging
    logging.basicConfig(
        format="%(asctime)s - %(levelname)s - %(name)s - %(message)s",
        datefmt="%m/%d/%Y %H:%M:%S",
        level=logging.INFO,
    )
    logger.info(accelerator.state, main_process_only=False)
    
    # Set seed
    if args.seed is not None:
        set_seed(args.seed)
    
    # Create output directory
    if accelerator.is_main_process:
        os.makedirs(args.output_dir, exist_ok=True)
    
    logger.info("=" * 70)
    logger.info("⚠️  FLUX INSTRUCTPIX2PIX TRAINING - CUSTOM IMPLEMENTATION")
    logger.info("=" * 70)
    logger.info("This is a custom training script for Flux InstructPix2Pix.")
    logger.info("Flux is very new - check for official implementations at:")
    logger.info("  - https://github.com/black-forest-labs/flux")
    logger.info("  - https://github.com/huggingface/diffusers")
    logger.info("")
    logger.info("Using LoRA for efficient fine-tuning...")
    logger.info("=" * 70)
    
    # Note: Full Flux implementation would go here
    # This includes:
    # 1. Loading Flux transformer model
    # 2. Setting up LoRA layers
    # 3. Loading dataset with proper preprocessing
    # 4. Training loop with flow matching objective
    # 5. Checkpoint saving
    
    logger.info("\n" + "=" * 70)
    logger.info("⚠️  IMPORTANT NOTE:")
    logger.info("=" * 70)
    logger.info("Flux is very new (released Aug 2024) and official")
    logger.info("InstructPix2Pix training scripts are not yet available.")
    logger.info("")
    logger.info("RECOMMENDED APPROACHES:")
    logger.info("")
    logger.info("1. ✅ Use SD 1.5 (proven, fast: ~1.5-2 days)")
    logger.info("   cd ../sd1_5 && ./train.sh --data_dir ./data_hf")
    logger.info("")
    logger.info("2. ✅ Use SDXL (better quality: ~3-5 days)")
    logger.info("   cd ../sdxl && ./train.sh --data_dir ./data_hf")
    logger.info("")
    logger.info("3. ⏰ Wait for official Flux InstructPix2Pix scripts")
    logger.info("   Monitor: https://github.com/huggingface/diffusers")
    logger.info("")
    logger.info("4. 🔧 Implement custom Flux training (advanced)")
    logger.info("   - Requires deep understanding of Flux architecture")
    logger.info("   - Flow matching training objective")
    logger.info("   - Proper handling of dual text encoders")
    logger.info("")
    logger.info("For now, we recommend starting with SDXL for best")
    logger.info("quality/time tradeoff, or SD 1.5 for fastest results.")
    logger.info("=" * 70)
    
    # For a production implementation, see README.md for resources
    
    accelerator.end_training()


if __name__ == "__main__":
    main()

